Added update methods to controllers

+++ chestnut/src/main/java/com/banking/chestnut/credit/repositories/CreditRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CreditRepository extends CrudRepository<Credits,Long> {
+public interface CreditRepository extends CrudRepository<Credits,Integer> {
 }
