+++ chestnut/src/main/java/com/banking/chestnut/credit/services/PaymentScheduleService.java
@@ -14,7 +14,7 @@ public class PaymentScheduleService {
     PaymentScheduleRepository paymentScheduleRepository;
 
 
-    public PaymentSchedule getById(Long id){
+    public PaymentSchedule getById(Integer id){
        return paymentScheduleRepository.findById(id).orElseThrow(NoSuchElementException::new);
     }
